access modifier moved to common
