+++ common/processor/class-model/src/main/java/io/helidon/common/processor/model/Field.java
@@ -20,6 +20,7 @@ import java.io.IOException;
 import java.util.Objects;
 import java.util.Set;
 
+import io.helidon.common.types.AccessModifier;
 import io.helidon.common.types.TypeName;
 
 /**
