+++ common/processor/class-model/src/main/java/io/helidon/common/processor/model/Method.java
@@ -26,6 +26,7 @@ import java.util.Set;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
+import io.helidon.common.types.AccessModifier;
 import io.helidon.common.types.TypeName;
 
 /**
