+++ inject/api/src/main/java/io/helidon/inject/api/ElementInfoBlueprint.java
@@ -20,6 +20,7 @@ import java.util.Optional;
 import java.util.Set;
 
 import io.helidon.builder.api.Prototype;
+import io.helidon.common.types.AccessModifier;
 import io.helidon.common.types.Annotation;
 import io.helidon.common.types.TypeName;
 import io.helidon.config.metadata.ConfiguredOption;
